movie_recommender_system/requirements.txt
# python 3.10
pandas>=1.5.2,<2.0.0
numpy>=1.23.5,<2.0.0
psycopg2>=2.9.6,<3.0.0
scikit-surprise>=1.1.3,<2.0.0
